pyproject.toml
[project]
name = "popmart-stock-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.54.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
popmart-stock-check = "main:main"